# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/properties/{id}/visit-templates": {
      "post": {
        "operationId": "VisitSlotsController_createTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a weekly visit-slot template",
        "tags": [
          "VisitSlots"
        ]
      },
      "get": {
        "operationId": "VisitSlotsController_listTemplates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List visit-slot templates for a property",
        "tags": [
          "VisitSlots"
        ]
      }
    },
    "/api/v1/visit-templates/{templateId}/deactivate": {
      "patch": {
        "operationId": "VisitSlotsController_deactivateTemplate",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate a template",
        "tags": [
          "VisitSlots"
        ]
      }
    },
    "/api/v1/properties/{id}/visit-slots/block": {
      "post": {
        "operationId": "VisitSlotsController_blockSlot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockSlotDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manually block a time range",
        "tags": [
          "VisitSlots"
        ]
      }
    },
    "/api/v1/properties/{id}/visit-slots": {
      "get": {
        "operationId": "VisitSlotsController_listAvailable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List available visit slots (public)",
        "tags": [
          "VisitSlots"
        ]
      }
    },
    "/api/v1/visits": {
      "post": {
        "operationId": "VisitsController_book",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookVisitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Book a visit slot",
        "tags": [
          "Visits"
        ]
      }
    },
    "/api/v1/visits/my": {
      "get": {
        "operationId": "VisitsController_myBookings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the user's own visit bookings",
        "tags": [
          "Visits"
        ]
      }
    },
    "/api/v1/visits/managed": {
      "get": {
        "operationId": "VisitsController_managedBookings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List visit bookings on managed properties",
        "tags": [
          "Visits"
        ]
      }
    },
    "/api/v1/visits/{id}/confirm": {
      "patch": {
        "operationId": "VisitsController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm a visit booking",
        "tags": [
          "Visits"
        ]
      }
    },
    "/api/v1/visits/{id}/cancel": {
      "patch": {
        "operationId": "VisitsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a visit booking",
        "tags": [
          "Visits"
        ]
      }
    },
...
    "/api/v1/sales/inquiries": {
      "post": {
        "operationId": "SalesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInquiryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a sale inquiry",
        "tags": [
          "Sales"
        ]
      }
    },
    "/api/v1/sales/inquiries/my": {
      "get": {
        "operationId": "SalesController_myInquiries",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the buyer's own inquiries",
        "tags": [
          "Sales"
        ]
      }
    },
    "/api/v1/sales/inquiries/managed": {
      "get": {
        "operationId": "SalesController_managed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List inquiries on managed properties",
        "tags": [
          "Sales"
        ]
      }
    },
    "/api/v1/sales/inquiries/{id}/status": {
      "patch": {
        "operationId": "SalesController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update inquiry status",
        "tags": [
          "Sales"
        ]
      }
    }
  },
  "info": {
    "title": "Paradis Immo API",
    "description": "Hybrid real estate platform API — Congo (CG). Generated snapshot.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTemplateDto": {
        "type": "object",
        "properties": {}
      },
      "BlockSlotDto": {
        "type": "object",
        "properties": {}
      },
      "BookVisitDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateInquiryDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateStatusDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}